requirements.txt
slackclient

python-dotenv

pytest
flake8